Fix crash in PSPSaveDir stream buffer offsets
Was trying to write into the `read` array starting at position 1000000 if it encountered a file larger than the buffer, causing it to enter the loop for a second time, having calculated the offset, rem, etc.

Streams manage their own Positions, we don't need to track offset. Offset is for the buffer in memory, and since we're starting from 0 each loop we don't need it, only need to know the bytesRead.

Should probably use safer FileStream APIs, but this fixes the issue for now.